requirements.txt
Flask==0.12.2
Flask-Mail==0.9.1
Jinja2==2.10
MarkupSafe==1.0
SQLAlchemy==1.2.5
Werkzeug==0.14.1
argparse==1.2.1
biopython==1.70
blinker==1.4
click==6.7
ete3==3.1.1
itsdangerous==0.24
numpy==1.14.2
redis==2.10.6
scipy==1.0.0
six==1.11.0
uWSGI==2.0.17
wheel==0.30.0
wsgiref==0.1.2
